+++ internal/caches/twitchusernamecache/username_loader.go
@@ -14,7 +14,7 @@ type UsernameLoader struct {
 	helixClient *helix.Client
 }
 
-func (l *UsernameLoader) Load(ctx context.Context, twitchUserID string, req *http.Request) ([]byte, time.Duration, error) {
+func (l *UsernameLoader) Load(ctx context.Context, twitchUserID string, req *http.Request) ([]byte, *int, *string, time.Duration, error) {
 	params := &helix.UsersParams{
 		IDs: []string{
 			twitchUserID,
@@ -23,14 +23,14 @@ func (l *UsernameLoader) Load(ctx context.Context, twitchUserID string, req *htt
 
 	response, err := l.helixClient.GetUsers(params)
 	if err != nil {
-		return nil, cache.NoSpecialDur, err
+		return nil, nil, nil, cache.NoSpecialDur, err
 	}
 
 	if len(response.Data.Users) != 1 {
-		return nil, cache.NoSpecialDur, errors.New("no user with this ID found")
+		return nil, nil, nil, cache.NoSpecialDur, errors.New("no user with this ID found")
 	}
 
 	user := response.Data.Users[0]
 
-	return []byte(user.Login), cache.NoSpecialDur, nil
+	return []byte(user.Login), nil, nil, cache.NoSpecialDur, nil
 }