+++ internal/resolvers/livestreamfails/clip_resolver.go
@@ -32,7 +32,7 @@ func (r *ClipResolver) Check(ctx context.Context, url *url.URL) (context.Context
 	return ctx, true
 }
 
-func (r *ClipResolver) Run(ctx context.Context, url *url.URL, req *http.Request) ([]byte, error) {
+func (r *ClipResolver) Run(ctx context.Context, url *url.URL, req *http.Request) (*cache.Response, error) {
 	clipID, err := clipIDFromContext(ctx)
 	if err != nil {
 		return nil, err